refactor: merge JwtValidationService into JwtService
